updated gui and reorganzied javafx files. cplan on converting homepage to a gridPane instead of anchorPane (maybe)
