+++ src/main/java/gui/App.java
@@ -1,0 +1,29 @@
+package gui;
+
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.stage.Stage;
+
+import java.util.Objects;
+
+public class App extends Application {
+
+    @Override
+    public void start(Stage primaryStage) throws Exception{
+
+        Parent root = FXMLLoader.load(Objects.requireNonNull(getClass().getResource("src/main/resources/fxml/homepage/HomePage.fxml")));
+        Scene scene = new Scene(root);
+
+
+
+        primaryStage.setTitle("Rotten Potatoes!");
+        primaryStage.setScene(scene);
+        primaryStage.show();
+    }
+ 
+ public static void main(String[] args) {
+        launch(args);
+    }
+}
