+++ src/main/java/gui/ContentController.java
@@ -1,19 +1,17 @@
-package javafx.src;
-
-import java.io.IOException;
-import java.util.Objects;
+package gui;
 
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.control.Button;
 import javafx.scene.control.TextField;
-import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.AnchorPane;
-import javafx.stage.Stage;
 
-public class HomePageController {
+import java.io.IOException;
+
+public class ContentController {
+
     @FXML
     private ImageView image1;
 
@@ -62,6 +60,7 @@ public class HomePageController {
 
     @FXML
     void loadSearchPage(ActionEvent event) throws IOException {
+        //System.out.println("search bar function");
         FXMLLoader loader = new FXMLLoader(getClass().getResource("/fxml/searchpage/SearchPage.fxml"));
