+++ src/main/java/gui/EpisodeOverviewController.java
@@ -1,7 +1,6 @@
-package javafx.src;
+package gui;
 
 import java.io.IOException;
-import java.util.Objects;
 
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
@@ -11,7 +10,7 @@ import javafx.scene.control.MenuItem;
 import javafx.scene.control.TextField;
 import javafx.scene.layout.AnchorPane;
 
-public class ShowOverviewController {
+public class EpisodeOverviewController {
 
     @FXML
     private Button homeButton;
@@ -23,25 +22,25 @@ public class ShowOverviewController {
     private TextField searchBar;
 
     @FXML
-    private MenuItem seasonSelector;
+    private TextField searchBar1;
 
     @FXML
+    private MenuItem seasonSelector;
+    
+    @FXML 
     private MenuItem episodeSelector;
 
-     @FXML
+    @FXML
     void loadHomePage(ActionEvent event) throws IOException {
+        FXMLLoader loader = new FXMLLoader(getClass().getResource("/fxml/homepage/HomePage.fxml"));
 
-         FXMLLoader loader = new FXMLLoader(getClass().getResource("/fxml/homepage/HomePage.fxml"));
-
-         AnchorPane pane = loader.load();
+        AnchorPane pane = loader.load();
         rootPane.getChildren().setAll(pane);
     }
 
     @FXML
     void loadSeasonPage(ActionEvent event) throws IOException {
-
         FXMLLoader loader = new FXMLLoader(getClass().getResource("/fxml/seasonoverview/SeasonOverview.fxml"));
-
 
         AnchorPane pane = loader.load();
         rootPane.getChildren().setAll(pane);
@@ -51,7 +50,9 @@ public class ShowOverviewController {
     void loadEpisodePage(ActionEvent event) throws IOException{
         FXMLLoader loader = new FXMLLoader(getClass().getResource("/fxml/episodeoverview/EpisodeOverview.fxml"));
 
+
         AnchorPane pane = loader.load();
         rootPane.getChildren().setAll(pane);
     }
+
 }
