+++ src/main/java/gui/SearchbarController.java
@@ -1,4 +1,4 @@
-package javafx.src;
+package gui;
 
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
