--- src/main/java/javafx/bin/HomePage.fxml
@@ -1,79 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.shape.Polygon?>
-<?import javafx.scene.shape.Rectangle?>
-<?import javafx.scene.text.Font?>
-
-<AnchorPane fx:id="rootPane" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="525.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/22" xmlns:fx="http://javafx.com/fxml/1" fx:controller="JavaFx.src.HomePageController">
-   <children>
-      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="WHITE" height="173.0" layoutX="33.0" layoutY="54.0" stroke="BLACK" strokeType="INSIDE" width="533.0" AnchorPane.bottomAnchor="298.0" AnchorPane.leftAnchor="33.5" AnchorPane.rightAnchor="33.5" AnchorPane.topAnchor="54.0" />
-      <Button fx:id="homeButton" layoutX="9.0" layoutY="9.0" mnemonicParsing="false" onAction="#loadHomePage" prefHeight="12.0" prefWidth="64.0" text="Home Page">
-         <font>
-            <Font size="10.0" />
-         </font>
-      </Button>
-      <TextField fx:id="searchBar" layoutX="84.0" layoutY="6.0" onAction="#loadSearchPage" prefHeight="26.0" prefWidth="392.0" promptText="Search">
-         <font>
-            <Font size="12.0" />
-         </font>
-      </TextField>
-      <Label layoutX="490.0" layoutY="13.0" text="Create Account | Login">
-         <font>
-            <Font size="10.0" />
-         </font>
-      </Label>
-      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="WHITE" height="113.0" layoutX="13.0" layoutY="255.0" stroke="BLACK" strokeType="INSIDE" width="575.0" />
-      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="#fcfcfc" height="113.0" layoutX="13.0" layoutY="396.0" stroke="BLACK" strokeType="INSIDE" width="575.0" />
-      <Button fx:id="showOverviewButton" layoutX="59.0" layoutY="197.0" mnemonicParsing="false" onAction="#loadShowOverview" text="load show overview">
-         <font>
-            <Font size="10.0" />
-         </font></Button>
-      <Label layoutX="22.0" layoutY="259.0" prefHeight="17.0" prefWidth="31.0" text="Genre">
-         <font>
-            <Font name="Arial Bold" size="10.0" />
-         </font>
-      </Label>
-      <Label layoutX="21.0" layoutY="400.0" prefHeight="17.0" prefWidth="31.0" text="Genre">
-         <font>
-            <Font name="Arial Bold" size="10.0" />
-         </font>
-      </Label>
-      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="DODGERBLUE" height="122.0" layoutX="59.0" layoutY="66.0" stroke="BLACK" strokeType="INSIDE" width="96.0" />
-      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="DODGERBLUE" height="122.0" layoutX="184.0" layoutY="66.0" stroke="BLACK" strokeType="INSIDE" width="96.0" />
-      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="DODGERBLUE" height="122.0" layoutX="313.0" layoutY="66.0" stroke="BLACK" strokeType="INSIDE" width="96.0" />
-      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="DODGERBLUE" height="122.0" layoutX="442.0" layoutY="66.0" stroke="BLACK" strokeType="INSIDE" width="96.0" />
-      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="DODGERBLUE" height="72.0" layoutX="56.0" layoutY="280.0" stroke="BLACK" strokeType="INSIDE" width="56.0" />
-      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="DODGERBLUE" height="72.0" layoutX="162.0" layoutY="280.0" stroke="BLACK" strokeType="INSIDE" width="56.0" />
-      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="DODGERBLUE" height="72.0" layoutX="272.0" layoutY="280.0" stroke="BLACK" strokeType="INSIDE" width="56.0" />
-      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="DODGERBLUE" height="72.0" layoutX="381.0" layoutY="280.0" stroke="BLACK" strokeType="INSIDE" width="56.0" />
-      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="DODGERBLUE" height="72.0" layoutX="482.0" layoutY="280.0" stroke="BLACK" strokeType="INSIDE" width="56.0" />
-      <Button fx:id="showOverviewButton1" layoutX="184.0" layoutY="197.0" mnemonicParsing="false" onAction="#loadShowOverview" text="load show overview">
-         <font>
-            <Font size="10.0" />
-         </font>
-      </Button>
-      <Button fx:id="showOverviewButton2" layoutX="313.0" layoutY="197.0" mnemonicParsing="false" onAction="#loadShowOverview" text="load show overview">
-         <font>
-            <Font size="10.0" />
-         </font>
-      </Button>
-      <Button fx:id="showOverviewButton3" layoutX="442.0" layoutY="197.0" mnemonicParsing="false" onAction="#loadShowOverview" text="load show overview">
-         <font>
-            <Font size="10.0" />
-         </font>
-      </Button>
-      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="DODGERBLUE" height="72.0" layoutX="56.0" layoutY="417.0" stroke="BLACK" strokeType="INSIDE" width="56.0" />
-      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="DODGERBLUE" height="72.0" layoutX="162.0" layoutY="417.0" stroke="BLACK" strokeType="INSIDE" width="56.0" />
-      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="DODGERBLUE" height="72.0" layoutX="272.0" layoutY="417.0" stroke="BLACK" strokeType="INSIDE" width="56.0" />
-      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="DODGERBLUE" height="72.0" layoutX="381.0" layoutY="417.0" stroke="BLACK" strokeType="INSIDE" width="56.0" />
-      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="DODGERBLUE" height="72.0" layoutX="482.0" layoutY="417.0" stroke="BLACK" strokeType="INSIDE" width="56.0" />
-      <Polygon layoutX="43.0" layoutY="348.0" points="-13.501102447509766, -33.0675048828125, -4.5, -57.70001220703125, -9.100000381469727, -57.70001220703125, -17.762968063354492, -33.0675048828125, -9.100000381469727, -8.5, -4.5, -8.5" stroke="BLACK" strokeType="INSIDE" />
-      <Polygon layoutX="46.0" layoutY="490.0" points="-13.501102447509766, -33.0675048828125, -4.5, -57.70001220703125, -9.100000381469727, -57.70001220703125, -17.762968063354492, -33.0675048828125, -9.100000381469727, -8.5, -4.5, -8.5" stroke="BLACK" strokeType="INSIDE" />
-      <Polygon layoutX="560.0" layoutY="351.0" points="4.4854736328125, -33.0675048828125, -4.5, -57.70001220703125, 1.08282470703125, -57.70001220703125, 10.10137939453125, -33.0675048828125, 1.08282470703125, -8.5, -4.5, -8.5" stroke="BLACK" strokeType="INSIDE" />
-      <Polygon layoutX="561.0" layoutY="488.0" points="4.4854736328125, -33.0675048828125, -4.5, -57.70001220703125, 1.08282470703125, -57.70001220703125, 10.10137939453125, -33.0675048828125, 1.08282470703125, -8.5, -4.5, -8.5" stroke="BLACK" strokeType="INSIDE" />
-   </children>
-</AnchorPane>
