--- src/main/java/javafx/bin/SearchPage.fxml
@@ -1,85 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.shape.Line?>
-<?import javafx.scene.shape.Rectangle?>
-<?import javafx.scene.text.Font?>
-
-<AnchorPane fx:id="rootPane" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/22" xmlns:fx="http://javafx.com/fxml/1" fx:controller="JavaFx.src.SearchPageController">
-   <children>
-      <TextField fx:id="searchBar" layoutX="84.0" layoutY="6.0" prefHeight="26.0" prefWidth="392.0" promptText="Search">
-         <font>
-            <Font size="12.0" />
-         </font>
-      </TextField>
-      <Button fx:id="homeButton" layoutX="9.0" layoutY="9.0" mnemonicParsing="false" onAction="#loadHomePage" prefHeight="12.0" prefWidth="64.0" text="Home Page">
-         <font>
-            <Font size="10.0" />
-         </font>
-      </Button>
-      <Label layoutX="490.0" layoutY="13.0" text="Create Account | Login">
-         <font>
-            <Font size="10.0" />
-         </font>
-      </Label>
-      <Label alignment="TOP_LEFT" layoutX="24.0" layoutY="68.0" prefHeight="16.0" prefWidth="102.0" text="Results for &quot;value&quot;" />
-      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="WHITE" height="103.0" layoutX="14.0" layoutY="145.0" stroke="BLACK" strokeType="INSIDE" width="367.0" />
-      <Label layoutX="118.0" layoutY="153.0" text="John Doe">
-         <font>
-            <Font name="Arial Bold" size="13.0" />
-         </font>
-      </Label>
-      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="WHITE" height="103.0" layoutX="14.0" layoutY="269.0" stroke="BLACK" strokeType="INSIDE" width="367.0" />
-      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="WHITE" height="103.0" layoutX="14.0" layoutY="393.0" stroke="BLACK" strokeType="INSIDE" width="367.0" />
-      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="#dadada" height="218.0" layoutX="415.0" layoutY="76.0" stroke="BLACK" strokeType="INSIDE" width="171.0" />
-      <Label alignment="TOP_LEFT" layoutX="430.0" layoutY="87.0" prefHeight="20.0" prefWidth="119.0" text="Filter" />
-      <VBox layoutX="443.0" layoutY="123.0" prefHeight="163.0" prefWidth="135.0" spacing="10.0">
-         <children>
-            <Label alignment="TOP_LEFT" prefHeight="20.0" prefWidth="119.0" text="Shows" />
-            <Label alignment="TOP_LEFT" prefHeight="20.0" prefWidth="119.0" text="Episodes" />
-            <Label alignment="TOP_LEFT" prefHeight="20.0" prefWidth="119.0" text="Studios" />
-            <Label alignment="TOP_LEFT" prefHeight="20.0" prefWidth="119.0" text="User's" />
-            <Label alignment="TOP_LEFT" prefHeight="20.0" prefWidth="119.0" text="Cast" />
-            <Label alignment="TOP_LEFT" prefHeight="20.0" prefWidth="119.0" text="Crew" />
-         </children>
-      </VBox>
-      <Line endX="96.0" layoutX="490.0" layoutY="107.0" startX="-75.0" />
-      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="#dadada" height="85.0" layoutX="24.0" layoutY="154.0" stroke="BLACK" strokeType="INSIDE" width="84.0" />
-      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="#dadada" height="85.0" layoutX="24.0" layoutY="278.0" stroke="BLACK" strokeType="INSIDE" width="84.0" />
-      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="#dadada" height="85.0" layoutX="24.0" layoutY="402.0" stroke="BLACK" strokeType="INSIDE" width="84.0" />
-      <Label layoutX="118.0" layoutY="278.0" text="&quot;Name of Show&quot;">
-         <font>
-            <Font name="Arial Bold" size="13.0" />
-         </font>
-      </Label>
-      <Label layoutX="118.0" layoutY="402.0" text="&quot;Episode Title&quot;">
-         <font>
-            <Font name="Arial Bold" size="13.0" />
-         </font>
-      </Label>
-      <Label alignment="TOP_LEFT" layoutX="118.0" layoutY="171.0" prefHeight="20.0" prefWidth="60.0" text="Actor" textFill="#757575">
-         <font>
-            <Font size="10.0" />
-         </font>
-      </Label>
-      <Label alignment="TOP_LEFT" layoutX="120.0" layoutY="298.0" prefHeight="20.0" prefWidth="162.0" text="Limited Series, 4 episodes, rating etc..." textFill="#757575">
-         <font>
-            <Font size="10.0" />
-         </font>
-      </Label>
-      <Label alignment="TOP_LEFT" layoutX="118.0" layoutY="423.0" prefHeight="20.0" prefWidth="222.0" text="Episode, &quot;TV Show name&quot; &quot;Season&quot; rating, etc.." textFill="#757575">
-         <font>
-            <Font size="10.0" />
-         </font>
-      </Label>
-      <Label alignment="TOP_LEFT" layoutX="24.0" layoutY="92.0" prefHeight="16.0" prefWidth="135.0" text="Did you mean &quot;______&quot;" textFill="#8d8c8c">
-         <font>
-            <Font name="Arial" size="10.0" />
-         </font>
-      </Label>
-   </children>
-</AnchorPane>
