+++ src/main/resources/fxml/content/Content.fxml
@@ -1,0 +1,56 @@
+<?import javafx.scene.layout.BorderPane?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.text.Font?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.layout.HBox?>
+
+
+<?import javafx.scene.layout.AnchorPane?>
+<BorderPane xmlns="http://javafx.com/javafx/8.0.65"
+            xmlns:fx="http://javafx.com/fxml/1"
+            fx:controller="gui.ContentController">
+
+
+    <!-- Top contains search bar, home button and login button-->
+    <top>
+
+        <HBox spacing="10" alignment="CENTER_LEFT" HBox.hgrow="ALWAYS" style="-fx-padding: 10">
+
+            <Button fx:id="homeButton" mnemonicParsing="false" onAction="#loadHomePage"
+                    prefHeight="30.0" text="Home Page">
+                <font>
+                    <Font size="10.0" />
+                </font>
+            </Button>
+
+            <TextField fx:id="searchBar" onAction="#loadSearchPage"
+                       prefHeight="26.0" HBox.hgrow="ALWAYS" promptText="Search">
+                <font>
+                    <Font size="12.0" />
+                </font>
+            </TextField>
+
+            <Label text="Create Account | Login">
+                <font>
+                    <Font size="10.0" />
+                </font>
+            </Label>
+
+        </HBox>
+
+    </top>
+
+
+    <center>
+
+        <HBox>
+            <AnchorPane fx:id="rootPane"/>
+            <fx:include source = "../homepage/HomePage.fxml" />
+        </HBox>
+
+    </center>
+
+
+
+</BorderPane>
