+++ src/main/resources/fxml/searchbar/Searchbar.fxml
@@ -5,7 +5,7 @@
 <HBox xmlns="http://javafx.com/javafx"
       xmlns:fx="http://javafx.com/fxml"
       fx:id="hBox"
-      fx:controller="javafx.src.SearchbarController">
+      fx:controller="gui.SearchbarController">
 
 <TextField fx:id="searchBar" layoutX="84.0" layoutY="6.0" onAction="#loadSearchPage" prefHeight="26.0" prefWidth="392.0" promptText="Search">
     <font>
